+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - `ProductDescription` content loader.
+- `Impersonate` component to  `Header`.
 
 ## [1.12.0] - 2018-08-02
 ### Added