+++ CHANGELOG.md
@@ -6,9 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Correct function from `orderFormContext` to add a new item to the orderForm.
 
 ## [1.12.2] - 2018-08-08
 ### Added
+- `ProductDescription` content loader.
 - `ProductName` and `ProductPrice` content loaders default style.
 
 ## [1.12.1] - 2018-08-08