+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.12.3] - 2018-08-09
 ### Fixed
 - Correct function from `orderFormContext` to add a new item to the orderForm.
 - `ProductPrice` condition now uses `isNil` function from `ramda`.